fix tests better
