+++ dev/conductor/ui/test/widgets/create_release_substeps_test.dart
@@ -6,22 +6,11 @@ import 'dart:async';
 
 import 'package:conductor_core/conductor_core.dart';
 import 'package:conductor_ui/widgets/create_release_substeps.dart';
-import 'package:file/file.dart';
-import 'package:file/memory.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:platform/platform.dart';
 
 import '../common.dart';
-
-/// A fake class to be mocked in order to test if nextStep is being called.
-class FakeNextStep {
-  void nextStep() {}
-}
-
-const String _flutterRoot = '/flutter';
-const String _checkoutsParentDirectory = '$_flutterRoot/dev/tools/';
 
 void main() {
   const String candidateBranch = 'flutter-1.2-candidate.3';
@@ -98,15 +87,12 @@ void main() {
   });
 
   group('The desktop app is connected with the CLI conductor', () {
-    const String exceptionMsg = 'There is a conductor Exception';
 
     testWidgets('Is able to display a conductor exception in the UI', (WidgetTester tester) async {
-      final FakeStartContext startContext = FakeStartContext();
-
-      startContext.addCommand(FakeCommand(
-        command: const <String>['git', 'clone', '--origin', 'upstream', '--', EngineRepository.defaultUpstream, '${_checkoutsParentDirectory}flutter_conductor_checkouts/engine'],
-        onRun: () => throw ConductorException(exceptionMsg),
-      ));
+      const String exceptionMsg = 'There is a conductor Exception';
+      final FakeStartContext startContext = FakeStartContext(
+        runOverride: () async => throw ConductorException(exceptionMsg),
+      );
 
       await tester.pumpWidget(
         StatefulBuilder(
@@ -137,13 +123,10 @@ void main() {
     });
 
     testWidgets('Is able to display a general exception in the UI', (WidgetTester tester) async {
-      final FakeStartContext startContext = FakeStartContext();
       const String exceptionMsg = 'There is a general Exception';
-
-      startContext.addCommand(FakeCommand(
-        command: const <String>['git', 'clone', '--origin', 'upstream', '--', EngineRepository.defaultUpstream, '${_checkoutsParentDirectory}flutter_conductor_checkouts/engine'],
-        onRun: () => throw Exception(exceptionMsg),
-      ));
+      final FakeStartContext startContext = FakeStartContext(
+        runOverride: () async => throw Exception(exceptionMsg),
+      );
 
       await tester.pumpWidget(
         StatefulBuilder(
@@ -175,9 +158,11 @@ void main() {
     });
 
     testWidgets('Proceeds to the next step if there is no exception', (WidgetTester tester) async {
-      final StartContext startContext = FakeStartContext();
-      final FakeNextStep fakeNextStep = FakeNextStep();
-
+      bool contextRunCalled = false;
+      bool nextStepReached = false;
+      final FakeStartContext startContext = FakeStartContext(
+        runOverride: () async => contextRunCalled = true,
+      );
       await tester.pumpWidget(
         StatefulBuilder(
           builder: (BuildContext context, StateSetter setState) {
@@ -186,7 +171,7 @@ void main() {
                 child: ListView(
                   children: <Widget>[
                     CreateReleaseSubsteps(
-                      nextStep: fakeNextStep.nextStep,
+                      nextStep: () => nextStepReached = true,
                       startContext: startContext,
                     ),
                   ],
@@ -203,6 +188,8 @@ void main() {
       await tester.pump();
       await tester.tap(continueButton);
       await tester.pumpAndSettle();
+      expect(contextRunCalled, true);
+      expect(nextStepReached, true);
     });
 
     testWidgets('Is able to display the loading UI, and hides it after release is done', (WidgetTester tester) async {
@@ -213,7 +200,7 @@ void main() {
       final Completer<void> completer = Completer<void>();
 
       startContext.addCommand(FakeCommand(
-        command: const <String>['git', 'clone', '--origin', 'upstream', '--', EngineRepository.defaultUpstream, '${_checkoutsParentDirectory}flutter_conductor_checkouts/engine'],
+        command: const <String>['git', 'clone', '--origin', 'upstream', '--', EngineRepository.defaultUpstream, '${kCheckoutsParentDirectory}flutter_conductor_checkouts/engine'],
         completer: completer,
       ));
 
@@ -251,101 +238,4 @@ void main() {
       expect(tester.widget<ElevatedButton>(continueButton).enabled, true);
     });
   });
-}
-
-class FakeStartContext extends StartContext {
-  factory FakeStartContext({
-    String candidateBranch = 'flutter-1.2-candidate.3',
-    Checkouts? checkouts,
-    String? dartRevision,
-    List<String> engineCherrypickRevisions = const <String>[],
-    String engineMirror = 'git@github:user/engine',
-    String engineUpstream = EngineRepository.defaultUpstream,
-    List<String> frameworkCherrypickRevisions = const <String>[],
-    String frameworkMirror = 'git@github:user/flutter',
-    String frameworkUpstream = FrameworkRepository.defaultUpstream,
-    Directory? flutterRoot,
-    String incrementLetter = 'm',
-    FakeProcessManager? processManager,
-    String releaseChannel = 'dev',
-    File? stateFile,
-    Stdio? stdio,
-  }) {
-    final FileSystem fileSystem = MemoryFileSystem.test();
-    flutterRoot ??= fileSystem.directory(_flutterRoot);
-    stateFile ??= fileSystem.file(kStateFileName);
-    final Platform platform = FakePlatform(
-      environment: <String, String>{'HOME': '/path/to/user/home'},
-      operatingSystem: const LocalPlatform().operatingSystem,
-      pathSeparator: r'/',
-    );
-    processManager = FakeProcessManager.list(<FakeCommand>[]);
-    stdio ??= TestStdio();
-    checkouts = Checkouts(
-      fileSystem: fileSystem,
-      parentDirectory: fileSystem.directory(_checkoutsParentDirectory),
-      platform: platform,
-      processManager: processManager,
-      stdio: stdio,
-    );
-    return FakeStartContext._(
-        candidateBranch: candidateBranch,
-        checkouts: checkouts,
-        dartRevision: dartRevision,
-        engineCherrypickRevisions: engineCherrypickRevisions,
-        engineMirror: engineMirror,
-        engineUpstream: engineUpstream,
-        flutterRoot: flutterRoot,
-        frameworkCherrypickRevisions: frameworkCherrypickRevisions,
-        frameworkMirror: frameworkMirror,
-        frameworkUpstream: frameworkUpstream,
-        incrementLetter: incrementLetter,
-        processManager: processManager,
-        releaseChannel: releaseChannel,
-        stateFile: stateFile,
-        stdio: stdio,
-    );
-  }
-
-  FakeStartContext._({
-    required String candidateBranch,
-    required Checkouts checkouts,
-    String? dartRevision,
-    required List<String> engineCherrypickRevisions,
-    required String engineMirror,
-    required String engineUpstream,
-    required List<String> frameworkCherrypickRevisions,
-    required String frameworkMirror,
-    required String frameworkUpstream,
-    required Directory flutterRoot,
-    required String incrementLetter,
-    required ProcessManager processManager,
-    required String releaseChannel,
-    required File stateFile,
-    required Stdio stdio,
-  }) : super(
-        candidateBranch: candidateBranch,
-        checkouts: checkouts,
-        dartRevision: dartRevision,
-        engineCherrypickRevisions: engineCherrypickRevisions,
-        engineMirror: engineMirror,
-        engineUpstream: engineUpstream,
-        flutterRoot: flutterRoot,
-        frameworkCherrypickRevisions: frameworkCherrypickRevisions,
-        frameworkMirror: frameworkMirror,
-        frameworkUpstream: frameworkUpstream,
-        incrementLetter: incrementLetter,
-        processManager: processManager,
-        releaseChannel: releaseChannel,
-        stateFile: stateFile,
-        stdio: stdio,
-  );
-
-  void addCommand(FakeCommand command) {
-    (checkouts.processManager as FakeProcessManager).addCommand(command);
-  }
-
-  void addCommands(List<FakeCommand> commands) {
-    (checkouts.processManager as FakeProcessManager).addCommands(commands);
-  }
 }
